requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
backports=1.1=pyhd3eb1b0_0
backports.functools_lru_cache=1.6.4=pyhd3eb1b0_0
backports_abc=0.5=py_1
blas=1.0=mkl
ca-certificates=2021.10.26=h06a4308_2
certifi=2020.6.20=pyhd3eb1b0_3
cycler=0.10.0=py27_0
dbus=1.13.18=hb2f20db_0
expat=2.2.10=he6710b0_2
fontconfig=2.13.0=h9420a91_0
freetype=2.10.4=h5ab3b9f_0
functools32=3.2.3.2=py27_1
futures=3.3.0=py27_0
glib=2.56.2=hd408876_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
icu=58.2=he6710b0_3
imutils=0.5.4=pypi_0
intel-openmp=2022.0.1=h06a4308_3633
jpeg=9b=habf39ab_1
kiwisolver=1.1.0=py27he6710b0_0
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libuuid=1.0.3=h1bed415_2
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
matplotlib=2.2.3=py27hb69df0a_0
mkl=2020.2=256
mkl-service=2.3.0=py27he904b0f_0
mkl_fft=1.0.15=py27ha843d7b_0
mkl_random=1.1.0=py27hd6b4f25_0
ncurses=6.3=h7f8727e_2
numpy=1.16.6=py27hbc911f0_0
numpy-base=1.16.6=py27hde5b4d6_0
opencv=2.4.11=nppy27_0
opencv-python=4.2.0.32=pypi_0
openssl=1.1.1m=h7f8727e_0
pcre=8.44=he6710b0_0
pillow=6.2.2=pypi_0
pip=19.3.1=py27_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py27h05f1152_2
python=2.7.18=ha1903f6_2
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2021.3=pyhd3eb1b0_0
qt=5.9.7=h5867ecd_1
readline=8.1.2=h7f8727e_1
setuptools=44.0.0=py27_0
singledispatch=3.7.0=pyhd3eb1b0_1001
sip=4.19.8=py27hf484d3e_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.37.0=hc218d9a_0
subprocess32=3.5.4=py27h7b6447c_0
tk=8.6.11=h1ccaba5_0
tornado=5.1.1=py27h7b6447c_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7f8727e_4
